front/requirements.txt
gradio
numpy
Pillow
tensorflow
